\name{casedata}
\alias{casedata}
\docType{data}
\title{Data set contains the number of cases information}
\description{
 Data set contains number of cases. i.e cancer.
}
\usage{data(casedata)}
\format{
data frame 
}
\details{
This is a data set contains 8 columns of cancer data set, 1999
  
}

\examples{
data(casedata)
## maybe str(casedata) ; plot(casedata) ...
}
\keyword{datasets}
