\name{excProb}
\alias{excProb}
 

\title{
 Exceedance probabilities
}
\description{
Calculate exceedance probabilities pr(X > threshold) from a fitted geostatistical model.}
\usage{
excProb(...)
}
 
\arguments{
\item{...}{Arguments passed to \code{\link[geostatsp]{excProb}} in the \code{geostatsp} package
}
}
\value{
results from \code{\link[geostatsp]{excProb}}
}
\seealso{
\code{\link[geostatsp]{excProb}}
\code{\link{bym}}
}





