\name{referencepop}
\alias{referencepop}
\docType{data}
\title{ Standard Canadian population data set from year 1991.}
\description{
  A data set contains population and age sex group from year 1991.
}
\usage{data(referencepop)}
\format{
  data frame
}
\details{
 This is a data set contains 36 rows and 3 columns. 
}

}
\examples{
data(referencepop)
## maybe str(referencepop) ; plot(referencepop) ...
}
\keyword{datasets}
