\name{popdata}
\alias{popdata}
\docType{data}
\title{Data set contains the information of population}
\description{
 Data set contains the information of population. It could be a shape file.
}
\usage{data(popdata)}
\format{
data in shape file
 
}
\details{
 It is a shape file. 
}


\examples{
data(popdata)
## maybe str(popdata) ; plot(popdata) ...
}
\keyword{datasets}
